pyproject.toml
[project]
name = "agent-tunix"
version = "0.1.0"
description = "GRPO Training for Gemma3-270m using Google Tunix"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["llm", "grpo", "reinforcement-learning", "gemma", "tunix", "jax"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-tunix[prod]>=0.1.3",
    "flax>=0.8.0",
    "jax>=0.8.0",
    "optax>=0.1.7",
    "orbax-checkpoint>=0.5.0",
    "grain>=0.1.0",
    "qwix>=0.1.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "tensorflow>=2.15.0",
    "tensorflow-datasets>=4.9.0",
    "tensorboardX>=2.6.0",
    "kagglehub>=0.2.0",
    "humanize>=4.8.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.20.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.20.0",
    "sphinx-notfound-page>=0.8.0",
]

[project.scripts]
agent-tunix = "agent_tunix.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/agent-tunix"
Repository = "https://github.com/yourusername/agent-tunix"
Documentation = "https://github.com/yourusername/agent-tunix#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_tunix"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
